First run at adding Threatstack agent
PLAT-70 #comment first pass at adding the threatstack agent to our
images

Fixing broken JSON

Moving threatstack into a wrapper class

Moving this into a wrapper due to how we are using hiera at the moment

Oh yeah, and the actual class too

:doh:

Moving threatstack from metadata to Puppetfile

Doing this because for testing at least I need to track a branch of our
fork of threatstack

Updated to correctly pull the right branch

Updated Puppetfile for testing

This tests the threatstack branch which fixes wget

Update deploy phose to start threatstack on boot

Deploy key isn't showing up in rc.local

PLAT-70 #comment the deploy key is nul in the file output... rebuilding
the image and trying again

Move threatstack into the deploy phase

PLAT-70 #comment moving threatstack install and config to the deploy
phase

Update your test!

Whoops.. travis failed because I told it to!

Removing hostname from the cloudsight setup

PLAT-70 #comment removing the --hostname argument, as it was picking up
the hostname of the deploy image host, not the hostname of the actual
deployed host. --hostname is optional anyway, so we should still get
the corrent hostname info

Hmmm trying to add FQDN back

PLAT-70 #comment we still aren't getting the nodes checking in with the
right hostname... grrr

Make sure that the configure_agent param is false

PLAT-70 #comment setting the configure agent parameter to false *in the
code* to make sure that the issue we are seeing isn't that hiera isn't
getting called, or populating the value for some reason.

Moving cloudsight setup to cloud-init

PLAT-70 #comment moving the first run of cloudsight set up from the
rc.local script, which seems to run when any run level changes, to the
cloud-init script, which sould only run when an instance starts